12-4 수업

+++ src/chap1_9/practice/q2/Employee.java
@@ -1,8 +1,8 @@
 package chap1_9.practice.q2;
 
 public class Employee {
-    private static int idCounter = 1;
-    private int employeeId;
+    private static int idCounter = 1; // 일련번호 누적을 관리
+    private long employeeId;
     private String name;
 
     public Employee(String name) {
